+++ packages/orq-rc/src/models/components/conversationresponse.ts
@@ -107,7 +107,7 @@ export const ConversationResponse$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kh6dher0kmk5tyzd2xt515m3"),
+  _id: z.string().default("conv_01kh6k3mqtbrf9br4e0bs6agfj"),
   kind: ConversationResponseKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),