+++ packages/orq-rc/src/models/operations/createconversationresponse.ts
@@ -280,7 +280,7 @@ export const PartReasoningPart$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("reasoning_01kh6dhesan4s2smj2f2mhq1g5"),
+  _id: z.string().default("reasoning_01kh6k3msjt0gc9jqve5yw300w"),
   metadata: z.record(z.any()).optional(),
   kind: PartKind$inboundSchema,
   reasoning: z.string(),
@@ -366,7 +366,7 @@ export const ReasoningPart$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("reasoning_01kh6dhert84vq78n47apr226j"),
+  _id: z.string().default("reasoning_01kh6k3msezkjse2sdf368ecrm"),
   metadata: z.record(z.any()).optional(),
   kind: DeltaKind$inboundSchema,
   reasoning: z.string(),