+++ packages/orq-rc/src/models/operations/filelist.ts
@@ -125,7 +125,7 @@ export const FileListData$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2026-02-11T13:16:27.706Z",
+    "2026-02-11T14:53:46.759Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {