+++ packages/orq-rc/src/models/operations/updateconversation.ts
@@ -226,7 +226,7 @@ export const UpdateConversationResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kh6dherdh2kwxcxa7zvxvbf4"),
+  _id: z.string().default("conv_01kh6k3mrsnbqskhfbh2datyq3"),
   kind: UpdateConversationKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),