+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-11T13:16:24.133Z",
+    "2026-02-11T14:53:43.170Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {